Add initial cut at sidewall computation addition
Issue Xyce/backlog/project-backlog#157 initially asked for the diode
to be augmented with the sidewall current computations that other
simulators (HSPICE, ngspice) support.

Garrick Ng had worked on doing this and had almost gotten it done, but
left before it was merged.  I found that his work was almost, but not
quite ready to merge, and embarked on an unrelated diode cleanup to
make it easier and cleaner to update.

This commit adds the first version of this additional computation.  I
have not tested the new computations, only that it compiles and does
not break any existing tests.

I have found a number of very minor errors in Garrick's original work,
and corrected them here.

I have also added the sidewall computations to the sensitivity code,
which Garrick had left for a future exercise.

I will get on with testing the new code now, and if there are no
further commits on this branch before I merge it, that'll be a sign
that all the new tests worked.

Xyce/backlog/project-backlog#157