Add output variables for Cd and Qd
In doing Xyce/ngspice comparisons, it was convenient to be able to
output the diode capacitance directly.  This commit adds diode
capacitance and diode charge to the output variables for the diode.
The would be accessible as N(instance:Cd) and N(instance:Qd) (replace
"instance" with the instance name).

Xyce/backlog/project-backlog#157