Make sure to use NBV where appropriate
The parameter NBV is used both in sidewall breakdown and normal diode
breakdown computation in other simulators.  Let's make sure to use it
in all places where it is supposed to be used.

I should note that Xyce has "supported" the NBV parameter for years,
but only in the sense that it didn't complain if you specified it.  It
basically ignored it ever since the supposedly "pspice-compatible"
level 2 diode was hacked in.

Now we actually use it in all diode levels.

Xyce/backlog/project-backlog#157  but only sorta.