Error check on number of args to random operators
This is related to this work:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/79
"Make "n" (AKA num_sigma) parameter optional for AGAUSS and GAUSS"

That issue is closed, but it occured to me today that I needed to make
it a little bit more robust.  This commit adds a check in each random
operator constructor to ensure that the user isn't attempting to pass
more than the maximum number of arguments.  This is a very small change,
but makes the code slightly safer.