Add brief explanation of sidewall computation to RG
The diode section of the RG has a section on the level 1 diode
equations that basically describe the SPICE3F5 diode with no
extensions.

I have added a section that documents the sidewall current and
capacitance equations, which are basically the same as the "bottom"
equations with different parameters.

Xyce/backlog/project-backlog#157