Allow expressions on analysis lines like .TRAN
"Add ability to parameterize transient end time"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/160

This commit fixes this issue, for the use case of constant expressions
in analysis statements such as .TRAN.

The more complex use case cases, of variable expressions will be the
subject of another issue.  This one is (mostly) done.

The two most important code changes for this issue were in the
"setValue" function in the N_UTL_Param.h file, and the
CircuitBlock::resolveExpressionsInOptionBlocks  function in the
N_IO_CircuitBlock.C file.

The "getValue" function is mostly used by various "setOptions" functions
to pull values from parameter objects and set them to local variables.
Prior to this change it exclusively used the "getImmutableValue"
function, which fails for most (all?) expressions.  I've conditionally
replaced it with "getMutableValue".

The resolveExpressionsInOptionBlocks function does exactly what the
name implies.  It performs all the necessary lookups, etc to resolve
dependencies in functions from various "dot" statements.  The main
change to this function to to increase the number of "dot" statements
that are considered in it.   This commit adds "TRAN", "DC", "AC",
and "HB" to the list.  Before, it was just "IC", "NODESET" and
"DOT_MEASURE_LINE".   Possibly in the future more options statements
will need to be added here.

One consequence of these changes is that the "getValue" function can no
longer reliably resolve non-expression string valued parameterss.  So,
this required some minor changes to the MOR::setMOROptions function and
also the new AnalysisManager::setDiagnosticMode function.