Update another setValue function for expressions
"Add ability to parameterize transient end time"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/160

In my previous commit on this issue, I updated the "setValue" function
in the N_UTL_Param.h file.  However, this function is overloaded and has
many versions.  For most of the versions, I want to leave them as-is,
but there was on that I overlooked that needed to be updated.

The one that I overlooked is identical the other one that I modified,
except that it also sets a "given" parameter.

Also, this commit removes a dreaded "#if 0" that I accidentically
included in my previous commit.