Add completely unmodified EKV 2.6 model source
This code cannot work in Xyce due to syntax issues, but is being
imported unmodified from its github repository
https://github.com/ekv26/model

This is the exact model as provided by its authors and released under
the Educational Community License.

Revisions to this model necessary to get it to work in Xyce/ADMS will
follow.

Xyce/backlog/project-backlog449