Make minimal changes to EKV so it will build
This commit makes the very minor changes necessary to the original EKV
2.6 source to allow it to build into a Xyce plugin.

This commit does NOT add the various attributes needed to get this
device plugged into Xyce as a MOSFET, only as a "Y" device.  Further
modifications are needed to get to that stage.  I am making this commit
as a separate step in order to show the process I had to go through to
resolve this story.

The changes are documented in README_Xyce

Xyce/backlog/project-backlog#449