Correct obvious error in TEMP range
The specified range of the TEMP variable is clearly wrong, as it says
it's supposed to be the temperature in degrees celsius, but has a
range specified as 273.15 to infinity.  This is clearly nonsense.

The TEMP variable is clearly used throughoug the device as a kelvin
temperature, and the parameter specified by the user in degrees C is later
converted to kelvin by adding 273.15 to it.

So clearly the desired range is -273.15 to infinity, not +273.15 to
infinity.

Fixing this eliminates a meaningless warning when reasonable celsius
temperatures are specified in the netlist.

Xyce/backlog/project-backlog#449