Refactor parsing of {schedule} on the tran line
This work is related to this issue:

"Add ability to parameterize transient end time"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/160

I recently modified the parser to allow expressions to be used for
arguments to analysis statements such as TRAN, DC, etc.  This mostly has
worked, but there was a complication on the 3rd and 4th arguments of the
TRAN line.

The complication is that for a long time we've allowed exactly one type
of expression on the TRAN line, using the "schedule" operator to allow
the maximum time step (DTMAX) to change during a simulation.  The way
the parser was written, if an expression is encountered after the first
2 required arguments, it is assumed to be DTMAX using "schedule".

This kind of conflicts with the way TRAN lines normally work, as DTMAX
is an optional fourth parameter (not 3rd).   I think the right way to
have done this would have been to require schedule-dependent
DTMAX parameter to be the fourth parameter on the line.  But for the
sake of backward compatibility I am preserving the old behavior.

Now, the parser allows all 4 arguments on the TRAN line to depend on
expressions.  If the 3rd or 4th line is time-dependent and/or uses the
"schedule" operator, it is assumed to be DTMAX (even if 3rd).
This works b/c the only parameter of the 4, for which it makes sense to allow
any sort of time dependence. is DTMAX.