Add lead current access for diagnostic output
Adding in both lead current access and the ability to
output to the diagnostic file all of the circuit nodes
that exceed a voltmax or current max.  This doesn't
work in parallel yet but does give a better picture
to the user about the status of the simulation.  Also,
I'm not sure what to do with internal device varaibles
that aren't of type 'v' or 'i' in the varType map

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/523