Add documentation for EKV 2.6
EKV 2.6 was in the "non free" repository as an undocumented and
basically unsupported device.  It has now been open sourced under the
ECL-2.0 license and has been moved to the open source repository for
Xyce.  It is now available to all users.

It is now documented insofar as a user can find reference to the level
260 mosfet and tables of its parameters, and also links to its
official documentation.

Addition of this model (or rather, its move to open source) is
documented in the 7.7 release notes.
Xyce/backlog/project-backlog#449