Fix DC statement to allow expressions
"Add ability to parameterize DC sweep arguments"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/536

This is also related to:
"Add ability to parameterize transient end time"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/160

The work on the second issue, above, contained some of the code changes
necessary to support the first issue.  But, not quite all.  The parsing
functions for DC statements relied a bit on the "isValue" function,
which only returns true if the string argument converts to a pure
number.  So, the function needed to be refactored a bit.

As part of this refactor,the handling of DATA statements in bot the
DC and STEP extract functions has been refactored.  This is a very
minor optimization.