Allow .TRAN line expressions to change via .STEP
This is related to:
"Add ability to parameterize transient end time"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/160

The originaal intent of this issue was to support static expressions
on the .TRAN line.  ie, expressions that cannot be changed after
parsing.  So, with my recent changes I was expecting that variable
expressions would not work.

I was correct in that expectation.  .TRAN arguments that depended on a
.param that was swept via .STEP did not properly update.  Each iteration
used the same initial value that was known at parse time.  I considered
adding code to have this condition emit an error.  But I then realized I
could just make variable expressions work correctly with a very small
amount of additional code.  That is contained in this commit.

Now, with today's changes, this will work correctly:

.param initial_time=0.0
.param tr_end=5ms
.tran {initial_time} {tr_end}
.step tr_end list 5ms 10ms 20ms

With each iteration, Xyce will perform a transient with a new value for
the final time.