Let configure detect ABI clashes
Ever since the very first commit of configure.ac (f5ebf77, back in
2002), after hunting down all required libraries, Xyce's configure
does a simple check to see if the entire link line actually works.
This was something that Tammy Kolda had suggested back in those days,
and I took her advice.

Thing is, when configure is hunting down libraries it's already trying
to link small programs with them, so this test has always been
pointless.  All it does is make sure the libraries can be found, which
honestly we already did earlier.

And worse, ever since we started using the Trilinos Makefile.export
thing instead of probing libraries normally, we NEVER actually link
meaningful test programs with trilinos anymore.

This means that while the libraries are found and a link of a trivial
program succeeds, it could be the case that Trilinos has been compiled
with a compiler that uses an ABI that is incompatible with the one
we're trying to use now.

This came up this weekend (again) when Eric tried to build Xyce on one
of the ASCIC machines and tried to use our precompiled Trilinos
archdir for gcc8.  These libraries were compiled with RHEL7 devtools-8
gcc 8.3, but Eric was using CDE gcc 8.4.

RHEL devtools gcc8 is compiled to use the old GCC ABI for compatibility
with the native gcc4, but CDE gcc 8.4 is compiled with the default,
new ABI.  It is not possible to use one with the other unless special
macros are defined to make it use the the non-default ABI.

And because we were only testing whether libraries would link by
linking an empty main function with the libraries, no Trilinos
functions are actually needed and so the linker never pulls them in
and tries to use them.

So configure succeeds, the build proceeds all the way to the final
link phase, and then a semi-infinite number of undefined symbole
errors show up.  It is frustrating and confusing for a user to have
that happen.

In this commit, I change this 20-year-old pointless link test into a
meaningful link test.  Now, configure tries to build a small test
program that includes Teuchos_RCP.hpp and tries to instantiate an
RCP.  This is a real test --- if the Trilinos libraries are
incompatible, this will fail, and fail before the user has bothered to
try to build Xyce.

if it fails, the user will get the message:
checking whether libraries (-lfftw3  -llocaepetra ...  ) will link... no
configure: error: FATAL: cannot link with the libraries detected so far ...
It may be that your compiler is incompatible with the one that was
used to build Trilinos.  See config.log for failure message.

That final error message used to include the entire list of libraries,
but that list is so freakin' long it makes the error message
unreadable (the beginning of the message scrolls off the screen before
the end of it gets printed).  So I shortened it.  The full list is in
the previous message anyway.

And config.log will have all of those undefined symbol errors instead
of them showing up at final link time of the full build.

This does not guarantee we won't get annoyed queries from users who
hit this problem, but it at least gets that frustration to
happen *before* they've spend an hour or so compiling Xyce.