Fix incorrect handling of probe-dependent integer variables
Xyce/ADMS emits code for derivatives of real variables --- it emits
declarations of variables to hold the derivatives, and carefully
propagates the derivatives through all expressions involving the
variables.

When an integer variable is assigned to with an RHS that has probes
in it, the variable is flagged as probe dependent and Xyce/ADMS was
half-trying to handle the derivatives with respect to probes.
Basically, it was not emitting declarations for the variables to hold
the derivatives, but sometimes would emit code that expected those
variables to exist.

Basically, as soon as you assign a probe-dependent expression to an
integer variable, you're effectively taking "int()" of that expression
and therefore creating something whose derivative is everywhere
zero *except* at discrete points where the expression is exactly an
integer (where the derivative is infinite).  Thus, the "correct" thing
to do is simply treat all such derivatives as if they were zero.

This came up for the first time when a user tried to have an integer
variable whose value was either +1 or -1 using a ternary operator with
a probe-dependent conditional.  But it has come up a couple of times
since in the google group in other, less trivial cases.

This commit fixes that.  Now it is possible to assign an integer
variable a probe-dependent expression, and Xyce/ADMS will generate
code that compiles.

Remember, though, that Xyce/ADMS always generates zero as the
derivative of such a variable with repect to any probe.

This commit also partially addresses a similar problem in analog
function context, but due to a very different issue there are still
serious problems with integer variables in analog functions.  This
issue was a blocker to fixing that one, so I wanted it out of the way.

Xyce/backlog/project-backlog#106