Fix vartype accessor function in parallel
When outputting diagnostic info in parallel the vartype function
which was modeled on the varname function wasn't working
correctly in parallel.  Also clean up the individual
processor based data gathering in parallel.  Off processor
results still need to be sent to the lead proc for output
in then diagnostic output file.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/523