Replace templated analog function with non-templated
In the initial implementation of Xyce/ADMS, I had handled all
differentiation of anything with respect to probes using Sacado.  As
part of this process, I had chosen to implement analog
functions (basically user-defined functions in Verilog-A) as templated
functions that had one typename, "ScalarT".  This way, the function
could be called with probe-independent arguments (which would be all
doubles) or probe-dependent arguments (Sacado FAD types), and the
compiler would just select which instantiation of the function to use.

When I jettisoned Sacado in favor of directly generated derivative
code (for efficiency), I initially followed the pattern of other open
source ADMS codes:  Have one function that just evaluates the analog
function, and another that just evaluates its derivative with respect
to one independent variable (which it does by evaluting the entire
function and its derivatives with respect to its arguments, and taking
a whole bunch of extra arguments that are the derivative of the
argument with respect to one independent variable --- the function
derivative is computed via the chain rule).

This, too, was extraordinarily inefficient for functions that have
many arguments and whose arguments depend on many independent
variables because the derivative function is even more complex than
the main function and has to be called repeatedly.

I ultimately jettisoned *that* in favor of a more complex but much,
much more efficient "evaluator" class that computes the function and
the derivatives with respect to arguments immediately upon
construction.  The value can then simply be retrieved from a stored
value, and the derivatives with respect to any probe are computed via
the chain rule with the stored partial derivatives with respect to
arguments --- so the function is called just once and the data reused.

But the evaluator classes are used only where "function
precomputation" is being used, which is only in the main module
context.  Function precomputation is NOT used inside analog functions
themselves, and so if any analog function calls any other analog
function, the old templated version *and* the derivative function are
still used.

This commit ONLY fixes the first of these problems, replacing the "all
arguments of the same type" template with a properly generated,
non-templated function where all the arguments have any type needed by
the user.

This commit is a necessary, but insufficient step towards addressing
issue Xyce/backlog/project-backlog#527.  It changes how *all* existing
models implement and use their analog functions, and so I had to
regenerate everything.  I did regenerate Xyce_NonFree, but the only
model in there uses no analog functions and so had no diffs other than
the generation date in the comment header.  I am not committing the
resulting regenerated NonFree code for that reason.

In order to address 527 completely, I still have to fix other broken
assumptions in the creation of the evaluator class and possibly the
derivative function.

Xyce/backlog/project-backlog#527