Make sure to initialize variables
The analog function derivative functions and the constructor of the
evaluator class were not careful to initialize temporary derivative
variables, and therefore could be a problem if nothing in the function
ever actually did anything to them.

This is the case when the variables are (incorrectly) declared as
real, but the function is computing all manner of integer
quantities (for which derivatives are NOT computed).

As it turns out, even though we're not strictly doing the correct
thing for functions that have integer return types, we're doing the
right arithmetic and returning the correct values (and ignoring
derivatives when they should be).  Making sure that variables are
correctly initialized lets this be close enough to correct that it'll
work in most cases (the rest of this issue still needs fixing, but
we're almost there).

This commit also changes the formerly templated "adms_vt"
function (which at the moment is generated for EVERY device in its own
namespace) into a static, non-templated function.  Using the templated
function definition was a holdover from the Sacado days, and there was
no case in which any use of this function *other* than with a double
argument was ever generated.  Now it's just a double function that
takes a double argument.

Xyce/backlog/project-backlog#527