Resolve parallel name look-up issue.
There was a problem when if the number of unknowns on each
processor was different, then there would be an unmatched
set of variable name look-ups.  This caused a problem even
when all of the look-ups were for locally owned unknowns because
the results of each variable name look-up was communicatd to all
processors.  This isn't needed as only the processor owning the
unknown needs to look up the variable name to generate a
diagnostic message.  Now each procesor can create a meaningful
diagnostic message given some diagnostic limits (maximum voltage
and/or maximum current).   The remaing problem is to get the
diagnostic messages to the lead processes.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/523