Augment html_params "set in" field
html_params.xml is an ADMS backend I wrote to examine the ADMS data
tree for a verilog-a model and produce an HTML page of that
information.

While digging into details of issue Xyce/backlog/project-backlog#529 I
found it helpful to check if a variable had been set in the
"@(initial_step)" block as well as all the other things it was
checking.  Now it does.

Xyce/backlog/project-backlog#529