Switch "initial_instance" use to "initial_step"
The diode2.va verilog-a module is an old diode model I wrote to
demonstrate use of "$limit" and to construct a toy DiodeClipper
netlist just like the one we use in the Users' Guide, but that ONLY
uses models defined in Verilog-A (no native devices at all, not even
voltage sources).

In investigating Xyce/backlog/project-backlog#529 a little, I
discovered that a long-standing claim from me that "@(initial_step)"
is very broken turned out to be an overstatment, and it works just
fine in some cases.  In fact, for this diode model it works great and
demonstrates how not completely broken it is.

So I'm switching this device module to use "@(initial_step)" to
initialize some bias-independent variables instead of
"@(initial_instance)".

This moves the generated code for that block OUT of "processParams"
and into "updateIntermediateVars" (inside a block that is only
executed in DCOP), but is in other ways just the same --- the
variables set here are still member variables of the instance class as
they need to be, and are retained in value across calls to
updateIntermediateVars.

As noted in Xyce/backlog/project-backlog#529, we do continue to get
reports from users about variable scoping issues when they use
"@(initial_step)" in ways *similar* to this, but at this point I am
unable to explain why --- what I *thought* was wrong is in fact not
wrong, so something else must be at the root of those problems.

The DiodeClipper that uses this diode2.va module is run every night as
part of the "Shared_Install" test jobs, so this is sort of a good
regression test to make sure that "@(initial_step)" does work at least
in some limited sense as intended.

Xyce/backlog/project-backlog#529