Very small order-of-operations change to support .IF
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/294
"Add support for .IF , .ELSE, .ELSEIF and .ENDIF"

This commit contains a very small change to the function
CircuitBlock::handleLinePass1.  This function is part of the "first pass" that
is peformed by the netlist parser and is called from both
CircuitBlock::parseIncludeFile and CircuitBlock::parseNetlistFilePass1.

The change is to the order in which .params/.globals are resolved vs. some of
the preliminary subcircuit processsing (to determine device counts and also
which subcircuits are purged).

As subcircuits (eventually) can be declared and/or used inside of a .IF
statement block, we don't know the full list of potential subcircuits and we
also don't know which ones are invoked until such .IF statements are resolved.

These .IF statements cannot be resolved until the param/global_params have
been resolved.  The previous order of operations was:

1. loop over all the lines, parse various things
2. do some preliminary subcircuit processing (including culling/purging)
3. then resolve the .param/.global_params.

Now it is:

1. loop over all the lines, parse various things
2. then resolve the .param/.global_params.
3.  --- placeholder for .IF expression resolution ----
4.  --- placeholder for .IF block parsing ----
5. do some preliminary subcircuit processing (including culling/purging)

(the placeholders for 3 and 4 are just in my mind at the moment)
I've run the tests in serial and this doesn't seem to break anything.
Hopefully this will hold up in regression testing.  It should, since the
subcircuit processing, above, doesn't depend on .params.