Adds interface for AC and transient analysis to optimization
The original ROL optimization code performed curve-fit optimization
for DC analysis.  This is accomplished through a lightweight interface
to the DCSweep object that enables the evaluation of objectives after
a successful sweep.  Similar infrastructure has been added to the AC
and transient analysis types so that ROL can integrate objectives and
constraints to enable an optimization for these analysis types.

The ROL netlist interface was extended to include these analyses types.
The AC and transient objectives need to be developed, as well as a
constraint object, so currently this code will just execute the simulation
without including it in the optimization.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/380