Fix for dashes in expr library TABLEFILE filenames
A user reported this today.  When using TABLEFILE with the expression
library, the expression parser didn't accept a filename that included a
dash.

This was very easy to fix. It required that the regex for the token
TOK_FILENAME be updated to include a dash. ("\-").

This issue is tracked here:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/560
"Expression support for TABLEFILE needs to allow the dash character in
filenames"