Allow TABLE/TABLEFILE data files have comments
This was requested by a user today.  Years ago I implemented this
feature in a device model, but never propagated it to the expression
library.  At the time, we were still using the old expresssion library.

Now that we have a new expression library, this is much easier to do.
The relevant function is nearly the same as the one from the device
model so this was mostly a cut-and-paste, and some minor edits.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/561
"Allow comments inside of data files read in via the TABLE/TABLEFILE
capability of the expression library"

With this commit, any characters to the right of '*', '$' or ';' will be
ignored.