Removes use of std::binary_function and std::unary_function
The binary_function and unary_function defined in the "functional"
header were deprecated in C++11 and removed in C++17.  Not that all
C++17 compilers will remove them, but they are free to do so.  Now
that Trilinos requires C++17, this may become an issue in the near
future, so it's worth removing them now.

It turns out that inheriting off binary_function and unary_function
only serve to define these members: result_type, first_argument_type,
and second_argument_type (only for binary).  These can be defined
directly in the structs that use binary_fuction and unary_function,
then the inheritence can be removed.  This has been done across the
code with this commit.