Changes call to std::remove_if to replace std::mem_fn with lambda
The std::mem_fn call used in the std::remove_if method is having issues
resolving arguments for a noexcept specifier, where exceptions cannot
be thrown by the function provided.  The simplest change I found to
address this problem for the compiler used in the TDEV testing is to
replace the std::mem_fn call with a lambda.  Lambdas are safe since
C++11, so this should be a robust change to the code.