Fix PWL source to work with .step/.param
Independent sources mostly work with .STEP, but PWL sources were
the exception.  This commit fixes the issue.  This issue is tracked
here:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/565
"Fix VPWL/IPWL independent sources to work with .STEP "

Prior to this commit the following voltage source would not update
correctly:

.param par1=2.0
.step par1 list 2.0 3.0 4.0
V1 A B PWL ( 0.0 1.0 0.1 'par1')

This code block specifies 3 .step iterations, with a different value for
par1 at each iteration.  However, prior to this commit, the PWL source
would use the initial value of par1, and never update.

Note, this equivalent Bsrc worked just fine, so there was a workaround:

.param par1=2.0
.step par1 list 2.0 3.0 4.0
B1 A B V={table(time,0.0,1.0,0.1,par1)}

Anyway, the reason the Vsrc version didn't work has to do with the
device package infastructure.  It was designed to conveniently update
parameter variables that are member variables of device instance and
device model classes.  Also, it is mostly designed to handle single
variables, not vectors.

Vsrcs (and Isrcs) don't really fit this design, since they rely on the
"SourceData" classes.  So, when .STEP is applied to a Vsrc parameter,
the device package infastructure updates a variable in the Vsrc instance
class, and then this Vsrc variable has to be awkwardly copied over
to the relevant SourceData class.  For the PWLinData class, (which is
that variant of SourceData used by the PWL source), the function the
performed this copy was never implemented.

The reason it was never implemented is because the PWL source involves
an array of arbitrary length, of time,value pairs.  The device parameter
infastructure isn't really set up for arrays.   But for this issue, I
managed to find a way to handle it.  It is still awkward, but it works.

Most of the changes in this commit are isolated to the PWLinData class
in the N_DEV_SourceData.h/C files.  So, those changes are pretty safe
and shouldn't break anything.

However, there is one small set of changes in the
N_DEV_DeviceEntity::setParams function, which potentially affect
many aspects of device model behavior in
Xyce.  I *think* this change will not have any bad impacts.  But,
if this commit causes problems, it will probably be from those changes.

The specific changes to the N_DEV_DeviceEntity::setParams is
the removal of 2 calls to param.setVal(number).  These didn't serve any
real purpose, except to effectively render any EXPR-type parameters as
constant (and thus make it impossible to obtain the underlying
expressiong, which I needed).  The param objects are only used during
setup, so this shouldn't matter much.

I've also  had to change a lot of "getImmutableValue" calls to
"getMutableValue" calls in the various SourceData classes for
related reasons.