clean up notifier/listener classes
This is related to this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/566
"Expression library breakpoint handling has a problem with .STEP"

.STEP makes use of the listener/notifier classes, which is a design
pattern for informing the various parts of Xyce that the simulation
needs to be reset for the next .STEP iteration.

There were a bunch of classes in this file, but I found that two of them
were not being used at all.  The unused classes were: ListenerProxy
and ListenerAdapter.  In this commit, I have removed them.

Also, the file contained a lot of comments, more comments than
actual code.  The comments didn't contain useful information
and were only there due to a previous developer's
misunderstanding our code standard.  So I have
removed most of these unnecessary comments as well.