Removal of expr library setupBreakpoints functions
This is related to this issue:

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/566
"Expression library breakpoint handling has a problem with .STEP"

This commit doesn't fix the issue, but is an interim "cleanup" commit.

As I dug through the code to try to fix this issue, I noticed a lot of
calls to the expression library's "setupBreakpoints" function.  However,
these functions are unconditionally no-ops.  So, they don't need to be
there and they don't need to be called.

I think they are a left over from an intermediate stage of developing
the expression libraary.  So, I created them a few years ago, but then
never needed to use them.  So, to avoid future confusion, I'm getting
rid of them.