Fix for bug in expression breakpointing with .STEP
Yesterday I discovered a small bug in the handling of breakpoints with
expressions.  This bug only happened with .STEP, on iterations after the
very first one.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/566
"Expression library breakpoint handling has a problem with .STEP"

The problem had to do with the order in which things were "reset" at the
top of the .STEP iteration.  Most of the code is reset via the various
"notify" functions that are called via the listener/notifier design
pattern.

The particular reset that was missing was resetting the "currTime"
variable in the device package early enough.  As the notifier code is
currently written, the notify functions are mostly called in the order
of class allocation.  So, the DeviceMgr::notify function is called
before the Transient::notify function.

Also, the notify functions were being called AFTER the relevant setParam
function was called.  When doing a .STEP, a parameter value is updated
at the beginning of each iteration via a call to DeviceMgr::setParam.
Prior to this commit,  the call to DeviceMgr::setParam was happening
before the notify calls.  It needs to happen after.

So, to summarize this commit changes the Analysis::Step::doLoopProcess
function so that the "notify" happens before the "updateSweepParams"
call, and the DeviceMgr::setParams function has been modified to include
updates to currTime, via a new function in the solverState class.  These
changes fix the problem.