Add a "downsample" option to expr library table op
This commit addresses this issue:

"Implement the ability to downsample data pulses"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/564

Many of the models in Xyce use the expression library to input data
files via the "table" and "tablefile" features.  However, it is often
the case that the data files are coming from experimental data sets and
can be very dense, much more dense than is necessary for simulation.

It is advantageous to sparsify this type of table, especially since the
expression library will (for most table variants) set breakpoints for
every single time point in a table.  For finely spaced time points, this
can cause significant slowdown in the time integrator.

I wrote the code for this a while ago, in a different context.  Sampling
a data set is pretty easy; one just has to interpolate.  However, the
problem is selection of points.  The approach I've taken in this commit
can be found in these papers:

 Yiming Zhao and Panagiotis Tsiotras.
 "Mesh Refinement Using Density Function for Solving Optimal Control Problems",
 AIAA Infotech@Aerospace Conference, Infotech@Aerospace Conferences, ()
 http:dx.doi.org/10.2514/6.2009-2019

 Yiming Zhao and Panagiotis Tsiotras.
 "Density Functions for Mesh Refinement in Numerical Optimal Control",
 Journal of Guidance, Control, and Dynamics, Vol. 34, No. 1 (2011), pp. 271-277.
 http:dx.doi.org/10.2514/1.45852

 The general idea is very simple:

 (1) Using the spline functions obtain the derivative array of the pulse file.
     Call it dfdt, where "f" is the values in the pulse file, and "t" is time.
 (2) Take absolute value of dfdt, so always >= 0.  Let dfdt be the "density function"
     to be used in refining the mesh.
 (3) Create a new array that is the integral of dfdt.  Kind of like a CDF.
     Call it F, where F(t_i) = \int_t_0^t_i dfdt dt
 (4) since dfdt is always >=0, the array/function F(t) is monotonically increasing.
 (5) Min of F is at F(t_0), Max of F is at F(t_N) where N=number of sample points.
 (6) Let the integral from F(t_i) to F(t_{i+1}) be a constant DF = (1/N)* F(t_N)
 (7) Set up a spline function, but with the purpose to provide F^{-1}.
     In other words, let the F array be "x" and the time array be "y".
     That way, for any given F, the spline will return time.
 (8) Loop i from 1 to N.   At each i, F = i*DF.  Use the F^{-1} spline to obtain time[i].
     At the end of the loop, the gradient-based time points will be set up.

This can be done with either the original pulse, or with the log10
of the original pulse.

This is invoked from the expression library via:

.param numPoints=100
.param logScale=1
table("filename.dat",numPoints,logScale)