Fixed an indexing error.
This is related to my work on:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/565
"Fix VPWL/IPWL independent sources to work with .STEP"

The work I did to fix this issue mostly addressed applying .STEP to
values in the PWL table.  However, I also set it up to potentially sweep
the TD and REPEATTIME parameters, which are separate from the table.  I
made a mistake in the REPEATTIME index.  This fixes it.