Changes error output for subcircuits during parser first pass
In the first pass of the parser, Xyce does store file and
circuit hierarchy that is contained in the CircuitContext
and CircuitBlock objects.  The name of subcircuits being
used is also collected during the first pass.  This allows
the parser to prune the hierarchical context table before
communicating it to other processors, in parallel.  Errors
and warnings output by the parser are provided along the way.

This commit changes the CircuitContext object to collect
error messages within the hierarchy that can be emitted as
warnings or errors depending upon whether a sub circuit is
pruned or kept after the first pass, respectively.  At a
minimum, this will address Ting's issues with empty,
unconnected subcircuits that have basic formatting errors.

However, the types of errors that can be collected is limited.
A netlist that defines a subcircuit that has no name will still
throw an error immediately.  Furthermore, it is not easy to
collect errors being generated by the CircuitBlock, which is
in charge of collecting all '.' lines like .MODEL, .PARAM, .FUNC,
etc.  These objects are immediately added to a non-hierarchical
list that checks whether they have been defined before and throws
an error if they have.  So, if a bad subcircuit also has a bad
'.' lines, that will not be ignored, even if the subcircuit is
not being used.

Xyce/backlog/project-backlog#555

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/555